+++ Source/Patches/EntranceTeleportPatches.cs
@@ -1,7 +1,5 @@
-using System.Linq;
 using HarmonyLib;
 using LCVR.Managers;
-using UnityEngine;
 
 namespace LCVR.Patches;
 
@@ -16,13 +14,8 @@ internal static class EntranceTeleportPatches
     [HarmonyPostfix]
     private static void OnTeleportPlayer(EntranceTeleport __instance)
     {
-        var doorPosition = Object.FindObjectsOfType<EntranceTeleport>().First(e =>
-                e.entranceId == __instance.entranceId && e.isEntranceToBuilding != __instance.isEntranceToBuilding)
-            .transform.position;
-        var direction = (__instance.exitPoint.position - doorPosition).normalized;
-        var angle = Vector3.SignedAngle(Vector3.forward, direction, Vector3.up) / 90;
-        var roundedAngle = (Mathf.Sign(angle) > 0 ? Mathf.Ceil(angle) : Mathf.Floor(angle)) * 90;
-        var rotation = roundedAngle - VRSession.Instance.MainCamera.transform.parent.localEulerAngles.y;
+        var entrancePoint = __instance.exitScript.entrancePoint;
+        var rotation = entrancePoint.eulerAngles.y - VRSession.Instance.MainCamera.transform.parent.localEulerAngles.y;
 
         VRSession.Instance.LocalPlayer.TurningProvider.SetOffset(rotation);
     }