+++ Source/Patches/HarmonyPatcher.cs
@@ -1,8 +1,6 @@
 ﻿using HarmonyLib;
 using System;
-using System.Linq;
 using System.Reflection;
-using UnityEngine;
 
 namespace LCVR.Patches;